pyproject.toml
[project]
name = "demoagents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "langsmith>=0.5.0",
    "lxml>=6.0.2",
    "openai>=2.14.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
]
